core/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
duckduckgo-search>=6.0.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
